requirements.txt
streamlit
spoon-ai-sdk
python-dotenv
requests
elevenlabs
networkx
pyvis
orjson
pydantic